requirements.txt
Flask==3.1.0
google-cloud-aiplatform==1.76.0
google-generativeai==0.8.4
gTTS==2.5.1
gunicorn==21.2.0
